Move codegen into core.
